+++ ui-library-call-readiness/src/components/PreCallChecksComponent.tsx
@@ -1,12 +1,14 @@
 import { useEffect, useState } from 'react';
-import { BrowserUnsupportedPrompt } from './UnsupportedBrowserPrompt';
+import { BrowserUnsupportedPrompt, BrowserVersionUnsupportedPrompt, OperatingSystemUnsupportedPrompt } from './UnsupportedEvironmentPrompts';
 import { CheckingDeviceAccessPrompt, PermissionsDeniedPrompt, AcceptDevicePermissionRequestPrompt } from './DevicePermissionPrompts';
 import { useCallClient } from '@azure/communication-react';
 import { checkBrowserSupport } from '../helpers/browserSupportUtils';
 import { checkDevicePermissionsState, requestCameraAndMicrophonePermissions } from '../helpers/devicePermissionUtils';
 
 type PreCallChecksState = 'runningChecks' |
+  'operatingSystemUnsupported' |
   'browserUnsupported' |
+  'browserVersionUnsupported' |
   'checkingDeviceAccess' |
   'promptingForDeviceAccess' |
   'deniedDeviceAccess' |
@@ -32,6 +34,7 @@ export const PreCallChecksComponent = (props: {
 
       // First we'll begin with a browser support check.
       const browserSupport = await checkBrowserSupport(callClient);
+      
       if (!browserSupport) {
         setCurrentCheckState('browserUnsupported');
         // If browser support fails, we'll stop here and display a modal to the user.
@@ -64,13 +67,17 @@ export const PreCallChecksComponent = (props: {
     };
 
     runCallReadinessChecks();
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   return (
     <>
+      <OperatingSystemUnsupportedPrompt isOpen={currentCheckState === 'operatingSystemUnsupported'}/>
       {/* We show this when the browser is unsupported */}
       <BrowserUnsupportedPrompt isOpen={currentCheckState === 'browserUnsupported'} />
+
+      {/* We show this when the browser version is unsupported */}
+      <BrowserVersionUnsupportedPrompt isOpen={currentCheckState === 'browserVersionUnsupported'} />
 
       {/* We show this when we are prompting the user to accept device permissions */}
       <AcceptDevicePermissionRequestPrompt isOpen={currentCheckState === 'promptingForDeviceAccess'} />
